some code cleanup and refactorization for test_mrsb case.
